% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepSelect.R
\name{StepSelect}
\alias{StepSelect}
\title{Stepwise significant block selection}
\usage{
StepSelect(z.block, alpha = 0.05)
}
\arguments{
\item{z.block}{a M * 2 matrix of z statistics}

\item{alpha}{thresholding parameter}
}
\value{
Return a list consist of two vectors:
1. A vector of p-value at each step  2.  A vector of identified significant block locations.
}
\description{
Stepwise significant block selection
}
\examples{

p <- 10^6; ## total number of pairs
X <- c(rep(0,p-30),rep(1,10),rep(2,10),rep(3,10));
X=0: no signal in either sequence of tests
X=1: signal in sequence 1 only
X=2: signal in sequence 2 only
X=3: simultaneous signal
set.seed(1);
Z1 <- rnorm(p,0,1); Z1[X==1|X==3] <- rnorm(20,3,1);
Z2 <- rnorm(p,0,1); Z2[X==2|X==3] <- rnorm(20,4,1);
StepSelect(cbind(Z1,Z2), alpha = 0.05 )
}
