% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Z.Block.R
\name{Z.Block}
\alias{Z.Block}
\title{Calculate the z statistics}
\usage{
Z.Block(X.1, X.2, outcome.1, outcome.2, same.ind = T)
}
\arguments{
\item{X.1}{genotypes matrix for outcome 1}

\item{X.2}{genotypes matrix for outcome 2}

\item{outcome.1}{a vecotr of outcome 1}

\item{outcome.2}{a vecotr of outcome 2}

\item{same.ind}{whether the collected outcome 1 and outcome 2 come from the same group of individuals}
}
\value{
a list of two blockwise z-statistics vectors
}
\description{
Calculate the z statistics
}
\examples{
data(test)
X.1 = data.list$X
Y.1 = data.list$Y1
Y.2 = data.list$Y2
Block.list = data.list$index
z.block <- matrix(NA, nrow = 1000, ncol = 2)
for(i in 1:1000){
  z.block[i,]  =  Z.Block(X.1 = X.1[, Block.list[[i]] ], X.2 = X.1[, Block.list[[i]] ],
                          outcome.1 = Y.1, outcome.2 =Y.2, )
}

}
